#27 [fix] rtrim in message.cpp
